Added fragment for note editor and added copyright to new classes
