+++ app/src/main/java/org/openstaxcollege/android/fragment/BookmarkFragment.java
@@ -1,3 +1,9 @@
+/**
+ * Copyright (c) 2015 Rice University
+ *
+ * This software is subject to the provisions of the GNU Lesser General
+ * Public License Version 2.1 (LGPL).  See LICENSE.txt for details.
+ */
 package org.openstaxcollege.android.fragment;
 
 import android.app.Activity;
