+++ app/src/main/res/layout/note_editor.xml
@@ -18,7 +18,7 @@ http://developer.android.com/resources/samples/NotePad/index.html
 		android:isScrollContainer="true"
 		android:saveEnabled="true" >
 		<view
-		    class="org.openstaxcollege.android.activity.NoteEditorActivity$LinedEditText"
+		    class="org.openstaxcollege.android.fragment.NoteEditorFragment$LinedEditText"
 		    android:id="@+id/note"
 		    android:layout_width="fill_parent"
 		    android:layout_height="fill_parent"
